+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 * Updated documentation page (#206)
 * Added `once` flag for scaffolding (#225)
-* Fixed potentially non-working Piral CLI scaffold on Windows (#192)
 * Added `defineDependency` to `piral-lazy` (#226)
+* Fixed potentially non-working Piral CLI scaffold on Windows (#192)
 * Fixed the use of `matchMedia` for other platforms (#222)
+* Fixed discarded props with multiple `includeProvider` calls (#227)
 
 ## 0.11.5 (May 24, 2020)